Keep information (address, data, control) stable until channel is ready
Enable the `core2axi` module to be used to connect the instruction
interface of the zero-riscy and ri5cy cores to the AXI bus.

According to the AXI specifiaction once address, data or control
information is presented on the bus by the source, this information
must be kept stable until the rising clock edge after the respective
`x_READY` signal was asserted.

This is already ensured in case the core interface keeps the address,
write data and byte enable information stable after asserting
`data_req_i` until `data_gnt_o` is received. For the the zero-riscy
and ri5cy cores this is the case for the LSU interface but not
for the instruction interface.

This is realized two major changes to the behavior of the module:
1. Always directly assert the `data_gnt_o` signal, even if the required
   AXI channel is not yet ready. This signals to the core that the
   address (and data) has been latched and that the transfer has to
   be finished
2. Keep the information (data, address, control) put on the AXI channels
   stable until the respective `x_READY` signal is sampled

Detailed changelog:
- Add registers for information presented on the core interface
  (write data, address, BE)
- Per default map core interface input address on the read channel
  and write channel address in the FSM.
- Introduce intermediate signals `data_be_int` and `data_wdata_int` to keep
  size adaption schemes for write data and write strobe AXI signals
  untouched (apart from using said intermediate signals instead of
  input signals).
- In the FSM per default map the respective input signals on the
  intermediate signals.
- In the IDLE state immediately assert `data_gnt_o` (combinatorial)
  once `data_req_i` is received
- Introduce two new states: `READ_ADDR` for waiting for the `ar_ready`
  signal and `WRITE_ADDR_DATA` for waiting for both the `aw_ready`
  and `b_ready` signal.
- After leaving the `IDLE` state keep the information put on the
  respective AXI channels stable until the respective `x_ready` signal
  is received.
- In the sequential part of the FSM constantly sample the address, data and
  be information presented on the core interface when in the `IDLE`,
  `ẀRITE_WAIT` or `READ_WAIT` states
- Remove direct mappings of core interface input signals to AXI output signals